GOT SONGS TO SHOW UP
I was assuming external storage would be to reference the sd card on an
Android device but I was wrong. Need to research a little more in
regards to storage systems in Android but I just switched my query for
the internal storage to external and I got songs to show up in my first
fragment. Now to get artists and genres working and on to saving user
selections.